refactor: update packs context
* Change exports
* Fetch actual proving process before a pack opening
* Fetch final proving process as soon as claim ends
