+++ js/packages/web/src/views/pack/components/OpenPackButtom/index.tsx
@@ -3,12 +3,12 @@ import { Button } from 'antd';
 import { usePack } from "../../contexts/PackContext";
 
 const OpenPackButton = ({ onClick }: { onClick: () => void }) => {
-  const { isProvingProcess } = usePack()
+  const { provingProcess } = usePack()
   return (
     <div className="open-pack">
       <div className="open-pack__title">Once opened a Pack cannot be resealed</div>
       <Button onClick={onClick}>
-        {isProvingProcess ? 'Resume Opening Pack' : 'Open Pack'}
+        {provingProcess ? 'Resume Opening Pack' : 'Open Pack'}
       </Button>
     </div>
   )
