+++ js/packages/web/src/views/pack/transactions/interface.ts
@@ -5,7 +5,7 @@ import BN from 'bn.js';
 import { PackSet } from '@oyster/common/dist/lib/models/packs/accounts/PackSet';
 
 import { ExtendedVoucherByKey } from '../../artworks/types';
-import { PackMetadataByPackCard } from '../hooks/useMetadataByPackCard';
+import { PackMetadataByPackCard } from '../contexts/hooks/useMetadataByPackCard';
 
 export interface GenerateTransactionsResponse {
   instructions: TransactionInstruction[];
